- Fixed IE related issues - [DONE] 183. FIX - page = nil related problem for WIllPagination - Added label for new field - phone - Fixed email notification related problem - [DONE] LS#12 Coming soon subscriber - [DONE] LS#13 Display list of subscribed users - [DONE] LS#2 Implement domain based virtual hosting support - [DONE] LS#11 Integrate front page featured products gallery - [DONE] LS#7 Create facebook and twitter fan pages - [DONE] LS#5 Implement Product gallery service - [DONE] LS#17 Add image zooming feature -

+++ app/controllers/products_controller.rb
@@ -1,6 +1,6 @@
 class ProductsController < ApplicationController
 
-  before_filter :login_required
+  before_filter :login_required, :only => [:new, :create, :edit, :update, :destroy]
   before_filter :detect_restaurant_and_product
 
   def index
@@ -8,7 +8,7 @@ class ProductsController < ApplicationController
     @products = @restaurant.products.image_attached.all(:limit => Product.per_page)
     @products_count = @restaurant.products.image_attached.count
 
-    if @restaurant.author?(current_user)
+    if logged_in? && @restaurant.author?(current_user)
       @products_for_admin = @restaurant.products.paginate(:page => params[:page])
     end
 
@@ -36,10 +36,20 @@ class ProductsController < ApplicationController
   end
 
   def edit
+    if @product && !@product.author?(current_user)
+      flash[:notice] = t('errors.authorization')
+      redirect_to site_products_url(@restaurant)
+    end
+
     render_view('products/edit')
   end
 
   def update
+    if @product && !@product.author?(current_user)
+      flash[:notice] = t('errors.authorization')
+      redirect_to site_products_url(@restaurant)
+    end
+
     product_params = params[:product]
     product_params.delete(:user_id)
     product_params.delete(:topic_id)
@@ -54,6 +64,11 @@ class ProductsController < ApplicationController
   end
 
   def destroy
+    if @product && !@product.author?(current_user)
+      flash[:notice] = t('errors.authorization')
+      redirect_to site_products_url(@restaurant)
+    end
+    
     if @product.destroy
       notify :success, "#{site_products_url(@restaurant)}#adminPortion"
     else
@@ -68,6 +83,29 @@ class ProductsController < ApplicationController
     end
   end
 
+  def show
+    render_view('products/show')
+  end
+
+  def slide
+    @first_image = @product.images.first
+    @images = @product.images
+    
+    render_view('products/slide')
+  end
+
+  def reviews
+    @site_title = 'Reviews'
+    @include_source_object = @product
+    @include_source_object_link = site_product_url(@product)
+    @attached_options = {
+        :attached_model => 'product',
+        :attached_id => @product.id
+    }
+    
+    render_view('reviews/index')
+  end
+
   private
     def detect_restaurant_and_product
       @product = nil
@@ -77,10 +115,6 @@ class ProductsController < ApplicationController
                     (@product ? @product.restaurant : nil) ||
                     Restaurant.find(params[:restaurant_id])
 
-      if @product && !@product.author?(current_user)
-        flash[:notice] = t('errors.authorization')
-        redirect_to site_products_url(@restaurant)
-      end
     end
 
     def find_from_post_data
